a1

+++ src/main/java/com/example/Restaurant/repository/AppetizerRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AppetizerRepository extends JpaRepository <Appetizer,Integer>{
+public interface AppetizerRepository extends JpaRepository <Appetizer, Long>{
 
     @Query("select (count(a) > 0) from Appetizer a where a.name = ?1")
     boolean existsByName(String name);
