+++ src/main/java/com/example/Restaurant/service/main/food/MainFoodServise.java
@@ -10,12 +10,12 @@ public interface MainFoodServise {
 
     void add(MainFood mainFood);
 
-    void deleteById(int id);
+    void deleteById(long id);
 
-    MainFood update(MainFood mainFood, int id);
+    MainFood update(MainFood mainFood, long id);
 
     List<MainFood> findall(int  pageNo ,int pageSize);
 
-    Optional<MainFood> findById(int id);
+    Optional<MainFood> findById(long id);
 
 }
